final push (fingers crossed)

+++ modules/sequelize.js
@@ -9,47 +9,47 @@
 
 const Sequelize = require("sequelize");
 
-// local database setup
-const sequelize = new Sequelize({
-  dialect: "sqlite",
-  storage: "database.sqlite",
-});
+let node_env = process.env.NODE_ENV;
+let sequelize;
 
-// live database setup
-// const sequelize = new Sequelize(process.env.JAWSDB_URL);
+if (node_env === "production") {
+  // live database setup
+  sequelize = new Sequelize(process.env.JAWSDB_URL);
+} else {
+  // local database setup
+  sequelize = new Sequelize({
+    dialect: "sqlite",
+    storage: "database.sqlite",
+  });
+  // local database reset
+  sequelize.sync({ force: true }).then(() => {
+    User.create({
+      firstName: "Paul",
+      lastName: "Heading",
+      emailAddress: "paulheading@me.com",
+      password: "$2a$10$BYqOQKj98RXZwLzrs6LaMOBa6/sHoKaLfYGbRhleaD1Lb/qrxN7v2",
+      level: "free",
+    });
+    Blacklist.create({
+      searchTerm: "grease",
+      albumId: "5n47Dui0H3pGpZSOxITmoq",
+      userId: "1",
+    });
+    Blacklist.create({
+      searchTerm: "grease",
+      albumId: "5WVPtycFFSm3uGhufMKwGn",
+      userId: "1",
+    });
+    Saved.create({
+      id: "1",
+      searchTerm: "grease",
+      userId: "1",
+    });
+  });
+}
 
-sequelize
-  .authenticate()
-  .then(() => {
-    console.log("Connection has been established successfully.");
-  })
-  .catch((err) => {
-    console.error("Unable to connect to the database:", err);
-  });
-
-sequelize.sync({ force: true }).then(() => {
-  User.create({
-    firstName: "Paul",
-    lastName: "Heading",
-    emailAddress: "paulheading@me.com",
-    password: "$2a$10$BYqOQKj98RXZwLzrs6LaMOBa6/sHoKaLfYGbRhleaD1Lb/qrxN7v2",
-    level: "free",
-  });
-  Blacklist.create({
-    searchTerm: "grease",
-    albumId: "5n47Dui0H3pGpZSOxITmoq",
-    userId: "1",
-  });
-  Blacklist.create({
-    searchTerm: "grease",
-    albumId: "5WVPtycFFSm3uGhufMKwGn",
-    userId: "1",
-  });
-  Saved.create({
-    id: "1",
-    searchTerm: "grease",
-    userId: "1",
-  });
+sequelize.authenticate().catch((err) => {
+  console.error("Unable to connect to the database:", err);
 });
 
 const User = sequelize.define("user", {
